Changes in the security file. Changes in the reporting path and in the rules path.
Changes in the location for the Rules actions (from the backend to the
RuleBundle called from the CalendarBundle).

Adding the Reporting view. Filters by month, by recruiter and by type.

Changes in the location of the filters. Now are located in the own view
of each one that needs to use a filter.