Changes in the calendar view. Now the appointment setter can see all the appointments for all the recruiters at the same time.
Changes in the jquery mobile config (hash listener disabled).

Changes in the notes. Now the appointment setter can see the notes of
all the recruiters at the same time.

Changes in the cancel button for some edit and new actions.

Changes in the filter. The recruiter is not required now.

Added asserts constraints to some entities.

Adding notes to the fixtures file.